Move response conversions to common
