+++ core/src/main/scala/org/genivi/sota/core/BlacklistResource.scala
@@ -30,7 +30,7 @@ class BlacklistResource(namespaceExtractor: Directive1[Namespace],
 
   import akka.http.scaladsl.server.Directives._
   import PackagesResource.extractPackageId
-  import org.genivi.sota.core.data.client.ResponseConversions._
+  import org.genivi.sota.rest.ResponseConversions._
   import org.genivi.sota.core.db.BlacklistedPackageResponse._
 
   implicit val _ec = system.dispatcher
