+++ core/src/main/scala/org/genivi/sota/core/data/client/PendingUpdateRequest.scala
@@ -15,6 +15,7 @@ import org.genivi.sota.core.data.UpdateStatus.UpdateStatus
 
 import scala.language.implicitConversions
 import org.genivi.sota.data.Interval
+import org.genivi.sota.rest.{GenericArgsDecoder, GenericResponseEncoder}
 import shapeless.{::, HNil}
 
 case class PendingUpdateRequest(requestId: UUID, packageId: PackageId, installPos: Int,
